+++ example/integration_test/plugin_integration_test.dart
@@ -10,16 +10,31 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'package:awty_engine_v2/awty_engine_v2.dart';
+import 'package:awty_engine_example/main.dart' as app;
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('getPlatformVersion test', (WidgetTester tester) async {
-    final AwtyEngineV2 plugin = AwtyEngineV2();
-    final String? version = await plugin.getPlatformVersion();
-    // The version string depends on the host platform running the test, so
-    // just assert that some non-empty string is returned.
-    expect(version?.isNotEmpty, true);
+  group('end-to-end test', () {
+    testWidgets('tap on the floating action button, verify counter',
+        (tester) async {
+      app.main();
+      await tester.pumpAndSettle();
+
+      // Verify the counter starts at 0.
+      expect(find.text('0'), findsOneWidget);
+
+      // Finds the floating action button to tap on.
+      final Finder fab = find.byTooltip('Increment');
+
+      // Emulate a tap on the floating action button.
+      await tester.tap(fab);
+
+      // Trigger a frame.
+      await tester.pumpAndSettle();
+
+      // Verify the counter increments by 1.
+      expect(find.text('1'), findsOneWidget);
+    });
   });
 }